requirements.txt
# Requirements for Flask
Flask==2.2.2
itsdangerous==2.1.2
jinja2==3.1.2
Werkzeug==2.2.2
click==8.1.3
python-dotenv==0.21.0	# dotenv
python-magic==0.4.27	# python magic
python-crontab==2.7.1	# cron
